+++ extensions/scryfall-search/src/CardDetails.tsx
@@ -1,5 +1,5 @@
 import { uid } from 'uid';
-import { List } from '@raycast/api';
+import { Detail } from '@raycast/api';
 
 import { MTGItem } from './typings';
 
@@ -10,7 +10,7 @@ interface CardDetailsProps {
 export const cardArt = (item: MTGItem, full: boolean) => {
   const dimensions = full
     ? 'raycast-width=420&raycast-height=585'
-    : 'raycast-width=336&raycast-height=468';
+    : 'raycast-width=252&raycast-height=351';
   const front = `![${item.name}](${item.image}?${dimensions})`;
   const back = item.back ? `![${item.name}](${item.back}?${dimensions})` : '';
   const markdown = `${front}${full ? '\n' : ''}${back}`;
@@ -19,54 +19,55 @@ export const cardArt = (item: MTGItem, full: boolean) => {
 };
 
 export const CardDetails = ({ item }: CardDetailsProps) => (
-  <List.Item.Detail
-    markdown={cardArt(item, false)}
+  <Detail
+    navigationTitle={item.name}
+    markdown={cardArt(item, true)}
     metadata={
-      <List.Item.Detail.Metadata>
-        <List.Item.Detail.Metadata.Label title="Name" text={item.name} />
+      <Detail.Metadata>
+        <Detail.Metadata.Label title="Name" text={item.name} />
         {item.colors && (
-          <List.Item.Detail.Metadata.TagList title="Mana Cost">
+          <Detail.Metadata.TagList title="Mana Cost">
             {item.colors?.map(item => (
-              <List.Item.Detail.Metadata.TagList.Item
+              <Detail.Metadata.TagList.Item
                 key={uid()}
                 text={item.text}
                 color={item.color}
                 icon={item.icon}
               />
             ))}
-          </List.Item.Detail.Metadata.TagList>
+          </Detail.Metadata.TagList>
         )}
-        <List.Item.Detail.Metadata.Label title="Type" text={item.type} />
-        <List.Item.Detail.Metadata.Separator />
+        <Detail.Metadata.Label title="Type" text={item.type} />
+        <Detail.Metadata.Separator />
         {item.oracle_text && (
-          <List.Item.Detail.Metadata.Label
+          <Detail.Metadata.Label
             title="Oracle Text"
             text={item.oracle_text}
           />
         )}
         {item.flavor_text && (
-          <List.Item.Detail.Metadata.Label
+          <Detail.Metadata.Label
             title="Flavor Text"
             text={item.flavor_text}
           />
         )}
         {(item.oracle_text || item.flavor_text) && (
-          <List.Item.Detail.Metadata.Separator />
+          <Detail.Metadata.Separator />
         )}
         {(item.power || item.toughness) && (
-          <List.Item.Detail.Metadata.Label
+          <Detail.Metadata.Label
             title="Power/Toughness"
             text={`${item.power}/${item.toughness}`}
           />
         )}
         {(item.power || item.toughness) && (
-          <List.Item.Detail.Metadata.Separator />
+          <Detail.Metadata.Separator />
         )}
-        <List.Item.Detail.Metadata.Label title="Rarity" text={item.rarity} />
-        <List.Item.Detail.Metadata.Label title="Set" text={item.set_name} />
-        <List.Item.Detail.Metadata.Separator />
-        <List.Item.Detail.Metadata.Label title="Artist" text={item.artist} />
-      </List.Item.Detail.Metadata>
+        <Detail.Metadata.Label title="Rarity" text={item.rarity} />
+        <Detail.Metadata.Label title="Set" text={item.set_name} />
+        <Detail.Metadata.Separator />
+        <Detail.Metadata.Label title="Artist" text={item.artist} />
+      </Detail.Metadata>
     }
   />
 );