+++ extensions/scryfall-search/CHANGELOG.md
@@ -8,3 +8,4 @@
 - Add list item details and action.
 - Add search list component and data quering.
 - Add extension example assets.
+- Update details to display full art and full details on full view.